api/requirements.txt
fastapi
uvicorn
rich
httpx
unidecode
ua-generator
yt-dlp
spotipy
python-multipart
python-dotenv
